requirements.txt
langchain
langchain-community
langchain-google-genai
google-generativeai
pydantic
faiss-cpu
chromadb
pypdf
python-dotenv
streamlit
neo4j
langgraph

